Move Logutil to Util dir
