+++ internal/proxy/proxy_test.go
@@ -98,9 +98,9 @@ import (
 	"github.com/milvus-io/milvus/internal/msgstream"
 
 	"github.com/milvus-io/milvus/internal/log"
-	"github.com/milvus-io/milvus/internal/logutil"
 	"github.com/milvus-io/milvus/internal/metrics"
 	"github.com/milvus-io/milvus/internal/rootcoord"
+	"github.com/milvus-io/milvus/internal/util/logutil"
 )
 
 const (
